extended FastAPI strawman
