+++ backend/database.py
@@ -1,0 +1,23 @@
+# database.py
+
+from sqlalchemy import create_engine, Column, Integer, String
+from sqlalchemy.ext.declarative import declarative_base
+from sqlalchemy.orm import sessionmaker
+
+DATABASE_URL = "postgresql://user:password@localhost:5432/mydatabase"
+
+Base = declarative_base()
+engine = create_engine(DATABASE_URL)
+SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
+
+class Order(Base):
+    __tablename__ = "orders"
+
+    id = Column(Integer, primary_key=True, index=True)
+    product_name = Column(String, index=True)
+    user_id = Column(String)
+    quantity = Column(Integer)
+
+# Create the database tables
+Base.metadata.create_all(bind=engine)
+
